+++ src/funcTest/groovy/info/solidsoft/gradle/pitest/functional/PitestPluginGradleVersionFunctionalSpec.groovy
@@ -3,10 +3,13 @@ package info.solidsoft.gradle.pitest.functional
 import com.google.common.base.Predicate
 import com.google.common.base.Predicates
 import groovy.util.logging.Slf4j
+import info.solidsoft.gradle.pitest.PitestPlugin
 import nebula.test.functional.ExecutionResult
 import nebula.test.functional.GradleRunner
+import org.gradle.api.GradleException
 import org.gradle.internal.jvm.Jvm
 import org.gradle.util.GradleVersion
+import spock.util.Exceptions
 
 import java.util.regex.Pattern
 
@@ -24,7 +27,7 @@ import static info.solidsoft.gradle.pitest.PitestTaskConfigurationSpec.PIT_PARAM
 class PitestPluginGradleVersionFunctionalSpec extends AbstractPitestFunctionalSpec {
 
     //4.8, but plugin requires 5.1.1
-    private static final GradleVersion MINIMAL_SUPPORTED_JAVA12_COMPATIBLE_GRADLE_VERSION = GradleVersion.version("5.1")
+    private static final GradleVersion MINIMAL_SUPPORTED_JAVA12_COMPATIBLE_GRADLE_VERSION = PitestPlugin.MINIMAL_SUPPORTED_GRADLE_VERSION
     //https://github.com/gradle/gradle/issues/8681#issuecomment-522951112
     private static final GradleVersion MINIMAL_SUPPORTED_JAVA13_COMPATIBLE_GRADLE_VERSION = GradleVersion.version("6.0-20190902220030+0000")
 
@@ -54,7 +57,27 @@ class PitestPluginGradleVersionFunctionalSpec extends AbstractPitestFunctionalSp
             requestedGradleVersion << applyJavaCompatibilityAdjustment(resolveRequestedGradleVersions()).unique()
     }
 
-    //To prevent failure when Spock for Groovy 2.4 is run with Groovy 2.3 delivered with Gradle <2.8
+    def "should fail with meaningful error message with too old Gradle version"() {
+        given:
+            gradleVersion = "5.0"
+        and:
+            assert PitestPlugin.MINIMAL_SUPPORTED_GRADLE_VERSION.compareTo(GradleVersion.version(gradleVersion)) > 0
+        when:
+            copyResources("testProjects/simple1", "")
+        then:
+            fileExists('build.gradle')
+        when:
+            ExecutionResult result = runTasksWithFailure('tasks')
+        then:
+            verifyAll {
+                def root = Exceptions.getRootCause(result.failure)
+                root.class.name == GradleException.name //name to mitigate differences on classloader
+                root.message.contains("'info.solidsoft.pitest' requires")
+                result.standardOutput.contains("WARNING. The 'info.solidsoft.pitest' plugin requires")
+            }
+    }
+
+//To prevent failure when Spock for Groovy 2.4 is run with Groovy 2.3 delivered with Gradle <2.8
     //Spock is not needed in this artificial project - just the test classpath leaks to Gradle instance started by Nebula
     private static final Pattern SPOCK_JAR_PATTERN = Pattern.compile(".*spock-core-1\\..*.jar")
     private static final Predicate<URL> FILTER_SPOCK_JAR = { URL url ->